requirements.txt
Django==1.10.5
django-filter==0.15.3
django-registration-redux==1.4
djangorestframework==3.5.3